Reduce table allocations in TSV output format
table.concat({...}, "\n") creates a table for each line, which becomes
immediate garbage. Handy when you don't know how many arguments will be
used, but here the number of arguments is known.  Suboptimal to allocate
more memory when this thing is usually run when a lot of memory is
already used.